# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: FUNHUB
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://funhub-mobile-backend.test'
paths:
...
  /api/v1/merchant/offers/my_bookmarks:
    get:
      summary: 'Get My Bookmarked Merchant Offers'
      operationId: getMyBookmarkedMerchantOffers
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {USER_TOKEN_KEY}'
          schema:
            type: string
      responses:
        200:
          description: success
          content:
            text/plain:
              schema:
                type: string
                example: "{\n \"data\": [],\n \"links\": {},\n \"meta\": {\n    \"current_page\": 1,\n  }\n}"
      tags:
        - Merchant
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                filter:
                  type: string
                  description: 'Column to Filter.'
                  example: 'Filterable columns are:  id, name, description, available_at, available_until, sku'
                filter_value:
                  type: string
                  description: 'Value to Filter.'
                  example: 'Filterable values are: 1, 2, 3, 4, 5, 6, 7, 8, 9, 10'
                sort:
                  type: string
                  description: 'Column to Sort.'
                  example: 'Sortable columns are: id, title, type, slug, status, published_at, created_at, updated_at'
                order:
                  type: string
                  description: 'Direction to Sort.'
                  example: 'Sortable directions are: asc, desc'
                limit:
                  type: integer
                  description: 'Per Page Limit Override.'
                  example: 10
                offset:
                  type: integer
                  description: 'Offset Override.'
                  example: 0
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []